Address comments.

+++ python/sglang/srt/lora/backend/chunked_backend.py
@@ -151,7 +151,7 @@ class ChunkedSgmvLoRABackend(BaseLoRABackend):
                     (num_segments,), dtype=torch.int32, device=self.device
                 ),
                 lora_ranks=torch.empty(
-                    (self.max_loras_per_batch,), dtype=torch.int64, device=self.device
+                    (self.max_loras_per_batch,), dtype=torch.int32, device=self.device
                 ),
                 scalings=torch.empty(
                     (self.max_loras_per_batch,), dtype=torch.float, device=self.device
